[CCUBE-1547][MAHI]Move mediaquery into breakpoint and use px value for spacing set
