+++ src/theme/spacing/specs/lifesg-spacing-set.ts
@@ -2,17 +2,17 @@ import { SpacingSet } from "../types";
 
 export const LifeSgSpacingSet: SpacingSet = {
     "spacing-0": 0,
-    "spacing-4": 0.25,
-    "spacing-8": 0.5,
-    "spacing-12": 0.75,
-    "spacing-16": 1,
-    "spacing-20": 1.25,
-    "spacing-24": 1.5,
-    "spacing-32": 2,
-    "spacing-40": 2.5,
-    "spacing-48": 3,
-    "spacing-64": 4,
-    "spacing-72": 4.5,
+    "spacing-4": 4,
+    "spacing-8": 8,
+    "spacing-12": 12,
+    "spacing-16": 16,
+    "spacing-20": 20,
+    "spacing-24": 24,
+    "spacing-32": 32,
+    "spacing-40": 40,
+    "spacing-48": 48,
+    "spacing-64": 64,
+    "spacing-72": 72,
 
     "layout-xs": 8,
     "layout-sm": 16,
